Whitespace: four spaces instead of one
Long ago I thought one space was plenty readable,
and if you are used to it, and look at it on the
right screen at the right distance with the size
and font good enough, and have good vision, yeah,
it is. But it is worse along several objective
dimensions of readability.